+++ src/SoftwareRender.h
@@ -3,6 +3,6 @@
 
 #include "gSP.h"
 
-f32 renderTriangles(const SPVertex * _pVertices, const u8 * _pElements, u32 _numElements);
+f32 renderTriangles(const SPVertex * _pVertices, const u16 * _pElements, u32 _numElements);
 
 #endif // SOFTWARE_RENDER_H
